# frontend/src/lib/api/swagger3.json
{
    "openapi": "3.0.0",
...
    "info": {
        "description": "API системы мониторинга репутации брендов BrandRadar",
        "title": "BrandRadar API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/alerts/config": {
            "post": {
                "description": "Создаёт конфигурацию spike-алерта для бренда",
                "tags": [
                    "Alerts"
                ],
                "summary": "Создать конфиг алерта",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.CreateAlertConfigRequest"
                            }
                        }
                    },
                    "description": "Конфигурация алерта",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.AlertConfigResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/alerts/config/{id}": {
            "get": {
                "description": "Возвращает конфигурацию алерта по ID",
                "tags": [
                    "Alerts"
                ],
                "summary": "Получить конфиг алерта",
                "parameters": [
                    {
                        "description": "UUID конфига",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.AlertConfigResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет параметры конфигурации алерта",
                "tags": [
                    "Alerts"
                ],
                "summary": "Обновить конфиг алерта",
                "parameters": [
                    {
                        "description": "UUID конфига",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.UpdateAlertConfigRequest"
                            }
                        }
                    },
                    "description": "Обновляемые поля",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.AlertConfigResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет конфигурацию алерта",
                "tags": [
                    "Alerts"
                ],
                "summary": "Удалить конфиг алерта",
                "parameters": [
                    {
                        "description": "UUID конфига",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/brands/{id}/alerts": {
            "get": {
                "description": "Возвращает историю сработавших алертов для бренда",
                "tags": [
                    "Alerts"
                ],
                "summary": "История алертов",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                                                        },
                                                        {
                                                            "type": "object",
                                                            "properties": {
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.AlertResponse"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/brands/{id}/alerts/config": {
            "get": {
                "description": "Возвращает все конфигурации алертов для бренда",
                "tags": [
                    "Alerts"
                ],
                "summary": "Конфиги алертов по бренду",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.AlertConfigResponse"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
        "schemas": {
            "prod-pobeda-2026_internal_controller_http_v1_dto.AlertConfigResponse": {
                "type": "object",
                "properties": {
                    "aggression": {
                        "type": "number",
                        "example": 0.5
                    },
                    "brand_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "cooldown_minutes": {
                        "type": "integer",
                        "example": 30
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "enabled": {
                        "type": "boolean",
                        "example": true
                    },
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "sentiment_filter": {
                        "type": "string",
                        "example": "negative"
                    },
                    "threshold": {
                        "type": "integer",
                        "example": 10
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "window_minutes": {
                        "type": "integer",
                        "example": 60
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.AlertResponse": {
                "type": "object",
                "properties": {
                    "brand_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "config_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "fired_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "mentions_count": {
                        "type": "integer",
                        "example": 15
                    },
                    "window_end": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "window_start": {
                        "type": "string",
                        "example": "2024-01-01T11:00:00Z"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.CreateAlertConfigRequest": {
                "type": "object",
                "required": [
                    "brand_id",
                    "cooldown_minutes",
                    "threshold",
                    "window_minutes"
                ],
                "properties": {
                    "aggression": {
                        "type": "number",
                        "maximum": 1,
                        "minimum": 0,
                        "example": 0.5
                    },
                    "brand_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "cooldown_minutes": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 30
                    },
                    "sentiment_filter": {
                        "type": "string",
                        "example": "negative"
                    },
                    "threshold": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 10
                    },
                    "window_minutes": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 60
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "message": {
                        "type": "string",
                        "example": "resource not found"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData": {
                "type": "object",
                "properties": {
                    "items": {},
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                    },
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.UpdateAlertConfigRequest": {
                "type": "object",
                "properties": {
                    "aggression": {
                        "type": "number",
                        "maximum": 1,
                        "minimum": 0,
                        "example": 0.5
                    },
                    "cooldown_minutes": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 30
                    },
                    "enabled": {
                        "type": "boolean",
                        "example": true
                    },
                    "sentiment_filter": {
                        "type": "string",
                        "example": "negative"
                    },
                    "threshold": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 15
                    },
                    "window_minutes": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 60
                    }
                }
            },
...
        }
    }
}